separate theme installation
