+++ packages/cli/README.md
@@ -11,6 +11,7 @@ npm i -g @vue-storefront/cli
 
 After installing CLI you'll get access to following commands:
 - `vsf init [dirname]` - Use it to set up new VS project. The CLI will ask you a few questions and set up new, ready to work with Vue Storefront instance.
-- `vsf generate-module [module-name]` - generates boilerplate for Vue Storefront module named `vsf-module-name` in a current directory
+- `vsf init:module [module-name]` - generates boilerplate for Vue Storefront module named `vsf-module-name` in a current directory
+- `vsf init:theme [dirname]` - generates theme for Vue Storefront, by default in a current directory
 - `vsf --help` - available commands
-- `vsf -version` - current version of cli
+- `vsf --version` - current version of cli
