+++ packages/cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@vue-storefront/cli",
-  "version": "0.0.16",
+  "version": "0.2.0",
   "description": "",
   "main": "index.js",
   "bin": {
