+++ packages/cli/scripts/install.js
@@ -6,10 +6,8 @@ const execa = require('execa')
 const spawn = require('child_process')
 const fs = require('fs')
 const semverSortDesc = require('semver/functions/rsort')
-const semverSatisfies = require('semver/functions/satisfies')
-const semverCoerce = require('semver/functions/coerce')
-const semverInc = require('semver/functions/inc')
-const merge = require('lodash/merge')
+const { options } = require('./../consts')
+const { createThemeTasks, createThemePrompt } = require('./../themeTasks')
 
 module.exports = function (installationDir) {
   installationDir = installationDir || 'vue-storefront'
@@ -18,36 +16,6 @@ module.exports = function (installationDir) {
     'develop',
     'master'
   ]
-
-  const options = {
-    version: {
-      stable: 'Stable version (recommended for production)',
-      rc: 'Release Candidate',
-      nightly: 'In development branch (could be unstable!)'
-    },
-    installation: {
-      installer: 'Installer (MacOS/Linux only)',
-      manual: 'Manual installation'
-    }
-  }
-
-  const themes = {
-    capybara: {
-      label: 'Capybara - based on Storefront UI',
-      branches: {
-        master: options.version.stable,
-        develop: options.version.nightly
-      },
-      minVsfVersion: '^1.11.0'
-    },
-    default: {
-      label: 'Default',
-      branches: {
-        master: options.version.stable
-      },
-      minVsfVersion: '*'
-    }
-  }
 
   const tasks = {
     installDeps: {
@@ -60,62 +28,7 @@ module.exports = function (installationDir) {
         return execa.shell(`git clone --quiet --single-branch --branch ${answers.specificVersion} https://github.com/DivanteLtd/vue-storefront.git ${installationDir} && cd ${installationDir}/core/scripts && git remote rm origin`)
       }
     },
-    cloneTheme: {
-      title: 'Copying Vue Storefront theme',
-      task: answers => execa.shell([
-        `git clone --quiet --single-branch --branch ${answers.themeBranch} https://github.com/DivanteLtd/vsf-${answers.themeName}.git ${installationDir}/src/themes/${answers.themeName}`,
-        `cd ${installationDir}/src/themes/${answers.themeName}`,
-        `git remote rm origin`
-      ].join(' && ')),
-      skip: answers => {
-        if (fs.existsSync(`${installationDir}/src/themes/${answers.themeName}`)) {
-          return `Chosen theme already exists in Vue Storefront installation directory ./${installationDir}/src/themes/`
-        }
-      }
-    },
-    configureTheme: {
-      title: 'Configuring Vue Storefront theme',
-      task: answers => {
-        const configurationFiles = ['local.config.js', 'local.json']
-        const [themeLocalConfigJsPath, themeLocalJsonPath] = configurationFiles.map(
-          file => `${installationDir}/src/themes/${answers.themeName}/${file}`
-        )
-        const vsfLocalJsonPath = `${installationDir}/config/local.json`
-        const vsfPackageJsonPath = `${installationDir}/package.json`
-
-        try {
-          const isVsfVersionAsBranch = ['master', 'develop'].includes(answers.specificVersion)
-          const vsfVersionFromPackageJson = JSON.parse(fs.readFileSync(vsfPackageJsonPath)).version
-          const vsfVersion = isVsfVersionAsBranch
-            ? semverInc(vsfVersionFromPackageJson, 'minor')
-            : vsfVersionFromPackageJson
-
-          const vsfLocalJson = fs.existsSync(vsfLocalJsonPath)
-            ? JSON.parse(fs.readFileSync(vsfLocalJsonPath))
-            : {}
-
-          const themeLocalJson = fs.existsSync(themeLocalConfigJsPath)
-            ? require(fs.realpathSync(themeLocalConfigJsPath))(vsfVersion)
-            : fs.existsSync(themeLocalJsonPath)
-              ? JSON.parse(fs.readFileSync(themeLocalJsonPath))
-              : null
-
-          if (themeLocalJson) {
-            fs.writeFileSync(vsfLocalJsonPath, JSON.stringify(merge(vsfLocalJson, themeLocalJson), null, 2))
-          }
-        } catch (e) {
-          console.error(`Problem with parsing or merging configurations (${configurationFiles})\n`, e)
-        }
-      },
-      skip: answers => {
-        const configurationFiles = ['local.config.js', 'local.json']
-        const themePath = `${installationDir}/src/themes/${answers.themeName}`
-
-        if (configurationFiles.every(file => !fs.existsSync(`${themePath}/${file}`))) {
-          return `Missing configuration file in theme folder (${configurationFiles}) - nothing to configure`
-        }
-      }
-    },
+    ...createThemeTasks(installationDir),
     runInstaller: {
       title: 'Running installer',
       task: () => spawn.execFileSync('yarn', ['installer'], { stdio: 'inherit', cwd: installationDir })
@@ -163,32 +76,7 @@ module.exports = function (installationDir) {
               return availableBranches
             }
           },
-          {
-            type: 'list',
-            name: 'themeName',
-            message: 'Select theme for Vue Storefront',
-            choices: answers => {
-              const isVsfVersionAsBranch = ['master', 'develop'].includes(answers.specificVersion)
-              const selectedVsfVersion = semverCoerce(answers.specificVersion)
-
-              return Object.entries(themes)
-                .filter(([, themeConfig]) => isVsfVersionAsBranch || semverSatisfies(selectedVsfVersion, themeConfig.minVsfVersion, { includePrerelease: true }))
-                .map(([themeName, themeConfig]) => ({
-                  name: themeConfig.label,
-                  value: themeName
-                }))
-            }
-          },
-          {
-            type: 'list',
-            name: 'themeBranch',
-            message: 'Select theme version',
-            choices: answers => Object.entries(themes[answers.themeName].branches)
-              .map(([branchName, branchLabel]) => ({
-                name: branchLabel,
-                value: branchName
-              }))
-          },
+          ...createThemePrompt(),
           {
             type: 'list',
             name: 'installation',
@@ -203,11 +91,11 @@ module.exports = function (installationDir) {
           const taskQueue = []
           taskQueue.push(tasks.cloneVersion)
           taskQueue.push(tasks.cloneTheme)
+          taskQueue.push(tasks.installDeps) // we need to install deps for theme
+          taskQueue.push(tasks.configureTheme)
           if (answers.installation === options.installation.installer) {
-            taskQueue.push(tasks.installDeps)
             taskQueue.push(tasks.runInstaller)
           }
-          taskQueue.push(tasks.configureTheme)
           new Listr(taskQueue).run(answers)
         })
     })
